+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Improve `inspectdb` adding support to `postgresql::numeric` data type
 - Add support for dynamically load DDL classes easing to add support to
   new databases without changing `Migrate` class logic
+- Fix decimal field change. (#246)
 
 ### 0.6.3